requirements.txt
playwright
playwright-stealth
beautifulsoup4
PyYAML
pymongo
underthesea
rank-bm25
